+++ KalikoCMS.Admin/Admin/Content/PropertyType/CollectionPropertyEditor.ascx.cs
@@ -53,7 +53,7 @@ namespace KalikoCMS.Admin.Content.PropertyType {
 
             ScriptManager.RegisterClientScriptInclude(this, typeof(CollectionPropertyEditor), "Admin.Content.PropertyType.CollectionPropertyEditor", "Content/PropertyType/CollectionPropertyEditor.js?v=" + Utils.VersionHash);
 
-            var clickScript = string.Format("top.propertyEditor.collection.openDialog('#{0}', null, '{1}');return false;", ListContainer.ClientID, _classParameter);
+            var clickScript = string.Format("top.propertyEditor.collection.openDialog($('#{0}'), null, '{1}');return false;", ListContainer.ClientID, _classParameter);
             AddNewButton.Attributes["onclick"] = clickScript;
 
             Items.Attributes.Add("data-type", _classParameter);