+++ ReleaseNotes.md
@@ -1,7 +1,9 @@
 ### New in 1.2.1
 * Method for setting default values for pages
 * Fixed problem where collection property editor breaks on larger property value
+* Fixed problem with non-responsive dialog when moving pages
 * Added IoC support for MVC controllers
+* Added option to select what fields to search in
 
 ### New in 1.2.0
 * Properties now recognize [Required]-attribute