+++ ReleaseNotes.md
@@ -3,6 +3,7 @@
 * Fixed problem where collection property editor breaks on larger property value
 * Fixed problem with non-responsive dialog when moving pages
 * Fixed problem with image editor in collection properties
+* Fixed problem with using composite properties in collections
 * Added IoC support for MVC controllers
 * Added option to select what fields to search in
 * Added PageMoved event